pyproject.toml
[project]
name = "exp-harness"
version = "0.1.0"
description = "A local experiments harness for reproducible multi-step runs."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "typer>=0.12",
]

[dependency-groups]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.6.0"]

[project.scripts]
run-experiment = "exp_harness.cli:app"
run_experiment = "exp_harness.cli:app"
run-expt = "exp_harness.cli:app"

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/exp_harness"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "docker: requires docker daemon and image availability",
  "slow: slow tests (concurrency/system)",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ARG"]
ignore = [
  "E501", # handled by ruff formatter
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
  "ARG001", # pytest fixtures and monkeypatch args
  "ARG002", # test doubles / protocol-style methods
  "ARG005", # lambdas used for monkeypatch
]
